src/requirements.txt
aiosqlite==0.21.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
coverage==7.7.0
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.9.0
joblib==1.4.2
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.4.0
pytz==2025.1
regex==2024.11.6
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.24
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0